refactor: ajusta imports e pacotes após reorganização
