+++ src/main/java/dev/java10x/GestaoDeProjetos/model/enums/CargoDoConsultor.java
@@ -1,4 +1,4 @@
-package dev.java10x.GestaoDeProjetos.Consultor;
+package dev.java10x.GestaoDeProjetos.model.enums;
 
 public enum CargoDoConsultor {
     DESENVOLVEDOR_JUNIOR,
