+++ src/main/java/dev/java10x/GestaoDeProjetos/repository/ConsultorRepository.java
@@ -1,5 +1,6 @@
-package dev.java10x.GestaoDeProjetos.Consultor;
+package dev.java10x.GestaoDeProjetos.repository;
 
+import dev.java10x.GestaoDeProjetos.model.ConsultorModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ConsultorRepository extends JpaRepository<ConsultorModel, Long> {
