+++ src/main/java/dev/java10x/GestaoDeProjetos/repository/ProjetoRepository.java
@@ -1,5 +1,6 @@
-package dev.java10x.GestaoDeProjetos.Projetos;
+package dev.java10x.GestaoDeProjetos.repository;
 
+import dev.java10x.GestaoDeProjetos.model.ProjetosModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProjetoRepository  extends JpaRepository<ProjetosModel, Long> {
